+++ src/SampSharp.Core/Callbacks/INewCallbackParameter.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace SampSharp.Core.Hosting;
+
+internal interface INewCallbackParameter
+{
+    object GetValue(IntPtr amx, IntPtr parameter);
+}
