+++ src/SampSharp.Core/GameModeBuilder.cs
@@ -15,11 +15,6 @@
 
 using System;
 using System.IO;
-using System.Linq;
-using System.Net;
-using System.Net.Sockets;
-using System.Reflection;
-using System.Runtime.InteropServices;
 using System.Text;
 using SampSharp.Core.CodePages;
 using SampSharp.Core.Logging;
