+++ src/SampSharp/SampSharp.def
@@ -6,8 +6,6 @@ EXPORTS
     AmxUnload
     ProcessTick
     OnPublicCall
-    sampsharp_print
-    sampsharp_register_callback
     sampsharp_fast_native_find
     sampsharp_fast_native_invoke
     sampsharp_get_api
