+++ src/SampSharp/hosted_server.cpp
@@ -20,7 +20,7 @@
 #define INTEROP_LIB "SampSharp.Core"
 #define INTEROP_CLASS INTEROP_LIB ".Hosting.Interop"
 
-hosted_server *hosting = NULL;
+hosted_server *hosting = nullptr;
 
 hosted_server::hosted_server(const char *clr_dir, const char* exe_path) {
     int retval;
@@ -31,23 +31,13 @@ hosted_server::hosted_server(const char *clr_dir, const char* exe_path) {
         log_error("Failed to initialize CoreCLR runtime. Error %d.", retval);
         return;
     }
-
-    if((retval = app_.create_delegate(INTEROP_LIB, INTEROP_CLASS, "Tick",
-        (void **)&tick_)) < 0) {
-        log_error("Failed to load Tick delegate. Error %d.", retval);
-    }
-    if((retval = app_.create_delegate(INTEROP_LIB, INTEROP_CLASS, "PublicCall",
-        (void **)&public_call_)) < 0) {
-        log_error("Failed to load PublicCall delegate. Error %d.", retval);
-    }
-
+    
     log_info("Starting game mode host...");
     hosting = this;
     const char *args[1];
     args[0] = "--hosted";
 
-    if((retval = app_.execute_assembly(sizeof(args) / sizeof(args[0]), args,
-        &exitcode)) < 0)  {
+    if((retval = app_.execute_assembly(std::size(args), args, &exitcode)) < 0)  {
         log_error("Failed to prepare game mode. Error %d.", retval);
         return;
     }
@@ -66,46 +56,5 @@ hosted_server::~hosted_server() {
 
     if(hosting == this) {
         hosting = NULL;
-    }
-}
-
-void hosted_server::tick() {
-    if(tick_) {
-        tick_();
-    }
-}
-
-void hosted_server::public_call(AMX *amx, const char *name, cell *params,
-    cell *retval) {
-    uint32_t 
-        response, 
-        len;
-
-    if(public_call_) {
-        len = LEN_CBBUF;
-        if(!callbacks_.fill_call_buffer(amx, name, params, buf_, &len, false)) {
-            return;
-        }
-
-        mutex_.lock();
-
-        response = public_call_(name, buf_, len);
-
-        mutex_.unlock();
-
-        if (retval) {
-            *retval = response;
-        }
-    }
-}
-
-void hosted_server::register_callback(uint8_t* buf) {
-    log_debug("Register callback %s", buf);
-    callbacks_.register_buffer(buf);
-}
-
-SAMPSHARP_EXPORT void SAMPSHARP_CALL sampsharp_register_callback(uint8_t *buf) {
-    if(hosting) {
-        hosting->register_callback(buf);
     }
 }
